requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
DAWG-Python==0.7.2
docopt==0.6.2
frozenlist==1.5.0
fuzzywuzzy==0.18.0
idna==3.10
joblib==1.4.2
Levenshtein==0.27.1
magic-filter==1.0.12
multidict==6.2.0
nltk==3.9.1
propcache==0.3.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dotenv==1.0.1
python-Levenshtein==0.27.1
RapidFuzz==3.12.2
regex==2024.11.6
rsa==4.9
Telethon==1.39.0
tqdm==4.67.1
typing_extensions==4.12.2
yarl==1.18.3
